pyproject.toml
[tool.poetry]
name = "pytrinets"
version = "0.1.0"
description = "Create and manipulate Petri nets in python!"
authors = ["AimbotParce <marcparcerisa@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
matplotlib = "^3.10.0"
gravis = "^0.1.0"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
